fix: 타입 파일 분리 및 타입 수정 (#61)
