+++ src/api/user.ts
@@ -1,5 +1,5 @@
-import { supabase } from 'shared/supabase';
-import { Typedata } from 'shared/supabase.type';
+import { supabase } from 'types/supabase';
+import { Typedata } from 'types/supabase.type';
 
 export const UserInfo = async (): Promise<Typedata['public']['Tables']['userinfo']['Row'][]> => {
   try {
