+++ src/components/Header/Header.tsx
@@ -6,17 +6,10 @@ import Tag from 'common/Tag';
 import CustomCarousel from 'common/CustomCarousel';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-
-interface mostPlayedGame {
-  appid: number;
-}
+import { GameType, GenreType } from 'types/games';
 
 interface HeaderProps {
-  mostPlayedGames: mostPlayedGame[];
-}
-
-export interface GenreType {
-  description: string;
+  mostPlayedGames: GameType[];
 }
 
 const Header = ({ mostPlayedGames }: HeaderProps) => {
@@ -67,7 +60,7 @@ const Header = ({ mostPlayedGames }: HeaderProps) => {
 
   // // 가장 많이 플레이된 게임 100개 중 top 5 appid 가져오기
   const topTen = mostPlayedGames?.slice(0, 5);
-  const appids = topTen?.map((game: mostPlayedGame) => game.appid) || [];
+  const appids = topTen?.map((game: GameType) => game.appid) || [];
 
   // top 5 상세 정보 가져오기
   const gameDetailsQueries = useQueries({
@@ -85,7 +78,6 @@ const Header = ({ mostPlayedGames }: HeaderProps) => {
 
   return (
     <StContainer>
-      {/* <StWrapper> */}
       <StCarouselWrapper>
         <CustomCarousel settings={settings}>
           {gameDetailsArray &&
@@ -96,7 +88,6 @@ const Header = ({ mostPlayedGames }: HeaderProps) => {
             ))}
         </CustomCarousel>
       </StCarouselWrapper>
-      {/* </StWrapper> */}
 
       {gameDetailsArray &&
         gameDetailsArray?.map((game, index) => (
