+++ src/pages/login/Login.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { supabase } from 'shared/supabase';
+import { supabase } from 'types/supabase';
 import { Link, useNavigate } from 'react-router-dom';
 import { StyledLogin, StyledForm, StyledInput, StyledButton, StyledH1, StyledLabel, StkakaoButton } from './styles';
 interface FormData {
