+++ src/pages/register/Register.tsx
@@ -28,7 +28,7 @@ import {
 import { ChangeEvent, useCallback, useEffect, useRef, useState } from 'react';
 import { useQuery, useMutation } from '@tanstack/react-query';
 import { useNavigate, useParams } from 'react-router-dom';
-import { supabase } from 'shared/supabase';
+import { supabase } from 'types/supabase';
 import { GENRE_NAME } from '../../constants/genre';
 import { getGames } from 'api/games';
 import { QUERY_KEYS } from 'query/keys';
